Compress the save data
Make objects support multiple Emoji to display because I want to use that eventually.
Store the display text as an index into the list of displays for the object since that takes up drastically less space than the actual character.
I've left in bit estimates in the save and load section because I want to eventually get to a system that compacts to a bit level.