Make NPCs Random But Persistent Gender and Race
Add NPCIdentifier class which creates an ID at initial room time references data in game save data so that skin tone and gender are consistent.
There was some nonsense with lazy static initialization so I've add an initialize function to InitialRooms so that the initial rooms objects are instantiated at a known time so that I know when to reset the npc counter. This implementation isn't great but it's the best I've got.
Tested with the vampire in the castle entryway.

Closes #4